Mathematics Mostly Done
The logic of the math behind filling in the bleachers is mostly complete.

There might be a couple bugs I still have to figure out with the change in X and Z for sim based on slope of length and slope of width of bleacher.  But it serves it's main function at the moment.

Still todo: Fetch random billboard png to go to the coords specified in the billboard row list, and output in a sim billboard format